fix: use latest error handling
