+++ packages/amplify-appsync-simulator/package.json
@@ -29,6 +29,8 @@
     "test": "jest --logHeapUsage --forceExit"
   },
   "dependencies": {
+    "amplify-cli-core": "3.2.2",
+    "amplify-prompts": "2.6.0",
     "@graphql-tools/schema": "^8.3.1",
     "@graphql-tools/utils": "^8.5.1",
     "amplify-velocity-template": "1.4.8",
