+++ packages/amplify-cli-core/src/errors/amplify-exception.ts
@@ -171,4 +171,5 @@ export type AmplifyFaultType =
   | 'TriggerUploadFault'
   | 'UnexpectedS3Fault'
   | 'UnknownFault'
-  | 'UnknownNodeJSFault';
+  | 'UnknownNodeJSFault'
+  | 'MockProcessFault';
